Added tests for stack trace parsing
