+++ packages/expo-cli/src/exp.ts
@@ -39,6 +39,7 @@ import { profileMethod } from './commands/utils/profileMethod';
 import Log from './log';
 import update from './update';
 import urlOpts from './urlOpts';
+import { matchFileNameOrURLFromStackTrace } from './utils/matchFileNameOrURLFromStackTrace';
 
 // We use require() to exclude package.json from TypeScript's analysis since it lives outside the
 // src directory and would change the directory structure of the emitted files under the build
@@ -378,27 +379,6 @@ Command.prototype.asyncAction = function (asyncFn: Action, skipUpdateCheck: bool
     }
   });
 };
-
-/**
- *
- * @param traceLine
- */
-function matchFileNameOrURLFromStackTrace(traceMessage: string): string | null {
-  if (!traceMessage.includes(' in ')) return null;
-  const traceLine = traceMessage.split(' in ')[0]?.trim();
-  // Is URL
-  // "http://127.0.0.1:19000/index.bundle?platform=ios&dev=true&hot=false&minify=false:110910:3 in global code"
-  if (traceLine.match(/https?:\/\//g)) {
-    const [url, params] = traceLine.split('?');
-
-    const paramsWithoutLocation = params.replace(/:(\d+)/g, '').trim();
-    return `${url}?${paramsWithoutLocation}`;
-  }
-
-  // "node_modules/react-native/Libraries/LogBox/LogBox.js:117:10 in registerWarning"
-  // "somn.js:1:0 in <global>"
-  return traceLine.replace(/:(\d+)/g, '').trim();
-}
 
 function getStringBetweenParens(value: string): string {
   const regExp = /\(([^)]+)\)/;
